Extract screenshot to HotwireViewScreenshot
